frontend_admin/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
azure-core==1.30.1
azure-identity==1.16.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.8.0
azure-keyvault-keys==4.9.0
azure-keyvault-secrets==4.8.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.6
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Principal==0.4.0
Flask-Security-Too==5.4.3
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_resources==6.4.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
msal==1.28.0
msal-extensions==1.1.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
passlib==1.7.4
portalocker==2.8.2
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wheel==0.41.2
WTForms==3.1.2
